Add third/party/rust/third_party.toml file
This file will declare all third-party crates that can be used from
first-party code. We will use it to find dependencies and generate
BUILD.gn files.

The `cc` crate is banned since it does C++ compilation for build.rs
build steps, and we will be writing those as GN rules instead, so we
add a README.md mentioning this. The `crates.py` tool will also
complain if the `cc` crate is attempted to be downloaded.

The third_party/PRESUBMIT.py file expects a third-party import to be
in the rust/ directory itself, which is incorrect. For now, we just
exclude the directory from checks. The `crates.py` tool generates
CHROMIUM.readme files and ensures the license is valid (or asks you
to specify the license). But we can add a pass to look into rust
crate dirs later.

Bug: 1069271
Change-Id: If7380f481065ad5e515e7972e1b7fbe553aa465f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3252442
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#936470}
NOKEYCHECK=True
GitOrigin-RevId: ea1d212d98e437b541af2368c48414d5f9d82224