[rust] Add autocxx and its dependencies
autocxx is an experimental Rust/C++ interoperability tool which
tries to combine the best features of cxx and bindgen. It's been
produced specifically with Chrome's needs in mind.

This CL adds autocxx and its remaining dependencies. Most of its
dependencies are already in the tree due to inclusion of bindgen
and cxx already.

They are populated in the usual way using crates.py download and
crates.py gen, with the following manual adjustments:

* Adding rustenv = [ "CARGO_PKG_VERSION=0.59.1" ] to
  autocxx_bindgen's BUILD.gn, in order to prevent bindgen
  depending on an external environment variable and thus rebuilding
  each time. (https://crbug.com/1296326)
* autocxx_gen/v0_13/BUILD.gn was essentially blank and has been
  populated. (https://crbug.com/1297719)
* Some duplicate build_rs_support targets were removed
  (https://crbug.com/1297722)
* We remove the bindgen executable from autocxx_bindgen, since
  we don't need it, and it conflicts with the same in the bindgen
  crate itself.
* This change includes a patch to proc-macro-error to work around
  a pinned dependency, which otherwise prevents correct generation
  of BUILD.gn (https://crbug.com/1296232)

Finally, autocxx is in intense flux all the time. This CL lands
0.16 but there have already been many pull requests since then,
and we should expect the code to change frequently. If this gets
to be bothersome, we should arrange for crates.py to become
aware of custom directories for specified crates, and then un-vendor
its code using a DEPS entry.

Bug: 1296155

Change-Id: I031a40efc8d4e68b1839e9f25eb6fbf99f4cc611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3321896
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Eric Foo <efoo@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978555}
NOKEYCHECK=True
GitOrigin-RevId: 28841c33c77833cc30b286f9ae24c97e7a8f4057