[rust] Upgrade to cxx 1.0.82
This upgrades cxx from 1.0.81 to 1.0.82, which is not a big upgrade; it
also upgrades cxxbridge_cmd to match. As we haven't used cxxbridge_cmd
for a while this is a more significant upgrade and requires pulling in a
new version of clap.

This cxxbrige_cmd upgrade may resolve a problem with Mac Rust builds,
though.

This CL:
* Downloads the new version (and new clap dependencies)
* Re-applies cxx patches
* Runs gnrt gen
* Updates inclusive_language exemption list to update clap version
  numbers

The gnrt generation here is subject to crbug.com/1393600 where some
deps lines are doubled. That doesn't seem to cause any problems, and
a CL in the near future will remove the problematic BUILD.gn anyway.

Bug: 1386212
Change-Id: I1e5ebab0079a8b233b9e83a96dfb439cda674c2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042584
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077658}
NOKEYCHECK=True
GitOrigin-RevId: 91644fccd5f143d1ed400d6e7ceea6e4c46c7d36