Add an audits.toml for cargo-vet, and add nom and minimal-lexical
Adds the criteria used by fuchsia and chromeos, which we can apply
to crates that we review. Hopefully in a way that cargo-vet reads
correctly.

`cargo vet` does not yet work, as it dies on the requirement for
`-Zbindeps` in Chromium (which we use to depend on cxxbridge-cmd).

The nom and minimal-lexical crates are reviewed and added to the
audits.toml to demonstrate examples.

Bug: 1314345, 1501342
Change-Id: Idef942268f99e40725caca74d8703ff50131e53e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5075486
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1232750}
NOKEYCHECK=True
GitOrigin-RevId: 10061fb94ae4f087de0faedb933a86064ea4c094