[Fontations] Roll Skrifa to 0.15.2
Require Skrifa 0.15.2 because of the workaround for non-random HashSet
initialization.  Required as a workaround for issue 1516634.

Minor versions of other crates updated in the process.

Followed instructions in [1], but passed "skrifa" argument to gnrt
update step to restrict the change to the Skrifa roll.

[1] https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/rust.md#updating-existing-third_party-crates

Bug: 1516634, 1516744
Change-Id: I06da321a8c06ad2e590b36ff3a4a0305facff0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5177611
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244567}
NOKEYCHECK=True
GitOrigin-RevId: 444e1e42d9ead6aa789bb12782d23947299da706