Encode the cxx version in one place: //build/rust/cxx_version.gni
Instead of writing the version in two places:
- The middle of //build/rust/BUILD.gn
- The cxx.h header file

Put it in a single gni file. We use that from BUILD.gn to get the path
to CXX, and set a define for the cxx.h header to use to include the
correct path.

This makes rolling cxx a little easier.

R=adetaylor@chromium.org, lukasza@chromium.org

Change-Id: I12f33c899d94ede35211721bbe633669ab8611a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5837611
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352755}
NOKEYCHECK=True
GitOrigin-RevId: 93b1a3d042cdca0b19fa993266e12f4aec924414