[owners] Remove palmer@chromium.org from third_party/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/palmer@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I786a262efae235519fa993f0b98b375ea1de9599
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554063
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1462219}
NOKEYCHECK=True
GitOrigin-RevId: 1a352f7b1f7bfd58a63a4fd3bf5ad3fc5522e5ea