Add xml Rust crate to Chromium Rust crates
Add the xml Rust cratex [1],[2] to Chromium as a future replacement for
libxml2 SAX parsing. Based on a working proof-of-concept [3] this crate
is the suitable replacement of choice. Written in safe Rust, and with a
streaming parsing API that is suitable for integration into
XMLDocumentParser.

A discussion of this crate vs. quick-xml can be found in the (internal) doc [5].

ATLs LGTM in: [4]

[1] https://crates.io/crates/xml
[2] https://github.com/kornelski/xml-rs
[3] https://chromium-review.googlesource.com/c/chromium/src/+/6828846
[4] https://groups.google.com/a/google.com/g/chrome-atls-discuss/c/ISFumR4PPeA/m/bNrxP3RQAgAJ
[5] http://go/chrome_x_mitigation

Bug: 441911594
Change-Id: I1c7a25756be5b016b314d3dba790c75c41338428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6959529
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1526334}
NOKEYCHECK=True
GitOrigin-RevId: 9cf3207424c8f73245666be413a7a13752dc32c4