[owners] Remove danieldilan@google.com from third_party/rust/png/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/danieldilan@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic3c9a9678f39e9b7d0965decf4238c766e56b7b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7132519
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1541858}
NOKEYCHECK=True
GitOrigin-RevId: 15256bada148acb91b8faaf30d808e36b443c13d